Former-commit-id: 74f7f4f923b8749d1da3f3fd98b38bbdf024497c

+++ Sources/Internal/UI/Private/Android/MovieViewControlAndroid.h
@@ -77,7 +77,7 @@ namespace DAVA
 class JniMovieViewControl
 {
 public:
-    JniMovieViewControl(uint32 id);
+    JniMovieViewControl(uintptr_t id);
     void Initialize(const Rect& rect);
     void Uninitialize();
 
@@ -93,7 +93,7 @@ public:
     bool IsPlaying() const;
 
 private:
-    uint32 id;
+    uintptr_t id;
 
     JNI::JavaClass jniMovieViewControl;
     Function<void(jint, jfloat, jfloat, jfloat, jfloat)> initialize;
